fix(stark-ui): ui:rename text-mask-new to text-mask
closed #2564
