--- packages/stark-ui/src/modules/input-mask-directives-new/directives/text-mask-new.constant.ts
@@ -1,5 +1,0 @@
-export class StarkTextMasksNew {
-	public static CREDITCARD_NUMBER = "0000-0000-0000-0000";
-
-	public static STRUCTURED_COMMUNICATION_NUMBER = "+++000/0000/00000+++";
-}
