+++ packages/stark-ui/src/modules/input-mask-directives/directives/email-mask.directive.ts
@@ -9,7 +9,7 @@ import { BooleanInput } from "@angular/cdk/coercion";
 /**
  * @ignore
  */
-const directiveName = "[starkEmailMaskNew]";
+const directiveName = "[starkEmailMask]";
 
 /**
  * @ignore
