+++ packages/stark-ui/src/modules/input-mask-directives/directives/number-mask-config.intf.ts
