+++ packages/stark-ui/src/modules/input-mask-directives/directives/number-mask.directive.spec.ts
@@ -4,27 +4,32 @@ import { FormControl, FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { By } from "@angular/platform-browser";
 import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { Observer } from "rxjs";
-import { BooleanInput } from "@angular/cdk/coercion";
-import { StarkEmailMaskNewDirective } from "./email-mask-new.directive";
+import { StarkNumberMaskDirective } from "./number-mask.directive";
+import { StarkNumberMaskConfig } from "./number-mask-config.intf";
 import { IMaskModule } from "angular-imask";
 
-describe("EmailMaskDirective", () => {
+describe("NumberMaskDirective", () => {
 	let fixture: ComponentFixture<TestComponent>;
 	let hostComponent: TestComponent;
 	let inputElement: DebugElement;
 
+	const numberMaskConfig: StarkNumberMaskConfig = {
+		prefix: "",
+		suffix: ""
+	};
+
 	@Component({
 		selector: "test-component",
-		template: getTemplate("[starkEmailMaskNew]='emailMaskConfig'")
+		template: getTemplate("[starkNumberMaskNew]='numberMaskConfig'")
 	})
 	class TestComponent {
-		public emailMaskConfig: BooleanInput = true;
+		public numberMaskConfig: StarkNumberMaskConfig = numberMaskConfig;
 		public ngModelValue = "";
 		public formControl = new FormControl("");
 	}
 
-	function getTemplate(emailMaskNewDirective: string): string {
-		return "<input " + "type='text' " + emailMaskNewDirective + ">";
+	function getTemplate(numberMaskDirective: string): string {
+		return "<input " + "type='text' " + numberMaskDirective + ">";
 	}
 
 	function initializeComponentFixture(): void {
@@ -48,7 +53,7 @@ describe("EmailMaskDirective", () => {
 	// Inject module dependencies
 	beforeEach(() => {
 		TestBed.configureTestingModule({
-			declarations: [StarkEmailMaskNewDirective, TestComponent],
+			declarations: [StarkNumberMaskDirective, TestComponent],
 			imports: [FormsModule, ReactiveFormsModule, IMaskModule],
 			providers: []
 		});
@@ -67,10 +72,10 @@ describe("EmailMaskDirective", () => {
 		});
 
 		it("should render the appropriate content", () => {
-			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkEmailMask directive
+			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkNumberMask directive
 		});
 
-		it("should update the input value and show the mask only when a valid event is triggered in the input field", () => {
+		xit("should update the input value and show the mask only when a valid event is triggered in the input field", () => {
 			// Angular2 text-mask directive handles only the "input" event
 			const validEvents: string[] = ["input"];
 
@@ -79,10 +84,10 @@ describe("EmailMaskDirective", () => {
 				fixture.detectChanges();
 				expect(inputElement.nativeElement.value).toBe("");
 
-				changeInputValue(inputElement, "my-email@", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
-				expect(inputElement.nativeElement.value).toBe("my-email@.");
+				expect(inputElement.nativeElement.value).toBe("12,345");
 			}
 
 			const invalidEvents: string[] = ["blur", "keyup", "change", "focus", "keydown", "keypress", "click"];
@@ -92,60 +97,73 @@ describe("EmailMaskDirective", () => {
 				fixture.detectChanges();
 				expect(inputElement.nativeElement.value).toBe("");
 
-				changeInputValue(inputElement, "my-email", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
-				expect(inputElement.nativeElement.value).toBe("my-email"); // no mask shown
+				expect(inputElement.nativeElement.value).toBe("12345"); // no mask shown
 			}
 		});
 
 		xit("should prevent invalid values to be entered in the input field when the value is changed manually", () => {
-			const invalidValues: string[] = ["@@", "@.a.", " @ .", "what@.ever@."];
+			const invalidValues: string[] = ["a", " ", "/*-+,."];
 
 			for (const value of invalidValues) {
 				changeInputValue(inputElement, value);
 				fixture.detectChanges();
 
-				expect(inputElement.nativeElement.value).toBe("@.");
+				expect(inputElement.nativeElement.value).toBe("");
+			}
+
+			const invalidNumericValues: string[] = ["1-2-3.4.5", "+1*23-4/5", ".1.234,5"];
+
+			for (const numericValue of invalidNumericValues) {
+				changeInputValue(inputElement, numericValue);
+				fixture.detectChanges();
+
+				expect(inputElement.nativeElement.value).toBe("12,345");
 			}
 		});
 
-		xit("should remove the mask only when the config is set to false", () => {
-			changeInputValue(inputElement, "my-email@");
+		xit("should refresh the mask whenever the configuration changes", () => {
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			expect(inputElement.nativeElement.value).toBe("my-email@.");
+			expect(inputElement.nativeElement.value).toBe("12,345");
 
-			hostComponent.emailMaskConfig = undefined;
+			hostComponent.numberMaskConfig = {
+				...numberMaskConfig,
+				prefix: "%",
+				suffix: " percent",
+				thousandsSeparatorSymbol: "-"
+			};
 			fixture.detectChanges();
 
-			changeInputValue(inputElement, "what@.ever@.");
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			expect(inputElement.nativeElement.value).toBe("my-email@."); // the mask is enabled by default
+			expect(inputElement.nativeElement.value).toBe("%12-345 percent");
+		});
 
-			hostComponent.emailMaskConfig = ""; // use case when the directive is used with no inputs: <input type='text' starkEmailMask>
+		it("should remove the mask when the config is undefined", () => {
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			changeInputValue(inputElement, "what@.ever@.");
+			expect(inputElement.nativeElement.value).toBe("12,345");
+
+			hostComponent.numberMaskConfig = <any>undefined;
 			fixture.detectChanges();
 
-			expect(inputElement.nativeElement.value).toBe("my-email@ ."); // the mask is enabled by default
-
-			hostComponent.emailMaskConfig = false;
+			changeInputValue(inputElement, "whatever+1*23-4/5");
 			fixture.detectChanges();
 
-			changeInputValue(inputElement, "what@@.ever@.");
-			fixture.detectChanges();
-
-			expect(inputElement.nativeElement.value).toBe("what@@.ever@."); // no mask at all
+			expect(inputElement.nativeElement.value).toBe("whatever+1*23-4/5"); // no mask at all
 		});
 	});
 
 	describe("with ngModel", () => {
 		beforeEach(
 			waitForAsync(() => {
-				const newTemplate: string = getTemplate("[(ngModel)]='ngModelValue' [starkEmailMaskNew]='emailMaskConfig'");
+				const newTemplate: string = getTemplate("[(ngModel)]='ngModelValue' [starkNumberMaskNew]='numberMaskConfig'");
 
 				TestBed.overrideTemplate(TestComponent, newTemplate);
 
@@ -159,7 +177,7 @@ describe("EmailMaskDirective", () => {
 		});
 
 		it("should render the appropriate content", () => {
-			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkEmailMask directive
+			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkNumberMask directive
 		});
 
 		xit("should update the input value and show the mask only when a valid event is triggered in the input field", () => {
@@ -171,10 +189,10 @@ describe("EmailMaskDirective", () => {
 				fixture.detectChanges();
 				expect(hostComponent.ngModelValue).toBe("");
 
-				changeInputValue(inputElement, "my-email@", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
-				expect(hostComponent.ngModelValue).toBe("my-email@ .");
+				expect(hostComponent.ngModelValue).toBe("12,345");
 			}
 
 			const invalidEvents: string[] = ["blur", "keyup", "change", "focus", "keydown", "keypress", "click"];
@@ -184,7 +202,7 @@ describe("EmailMaskDirective", () => {
 				fixture.detectChanges();
 				expect(hostComponent.ngModelValue).toBe("");
 
-				changeInputValue(inputElement, "my-email@", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
 				// IMPORTANT: the ngModel is not changed with invalid events, just with "input" events
@@ -193,7 +211,7 @@ describe("EmailMaskDirective", () => {
 		});
 
 		xit("should prevent invalid values to be entered in the input field when the value is changed manually", () => {
-			const invalidValues: string[] = ["@@", "@.a.", " @ .", "what@.ever@."];
+			const invalidValues: string[] = ["a", " ", "/*-+,."];
 
 			for (const value of invalidValues) {
 				changeInputValue(inputElement, value);
@@ -201,37 +219,48 @@ describe("EmailMaskDirective", () => {
 
 				expect(hostComponent.ngModelValue).toBe("");
 			}
+
+			const invalidNumericValues: string[] = ["1-2-3.4.5", "+1*23-4/5", ".1.234,5"];
+
+			for (const numericValue of invalidNumericValues) {
+				changeInputValue(inputElement, numericValue);
+				fixture.detectChanges();
+
+				expect(hostComponent.ngModelValue).toBe("12,345");
+			}
 		});
 
-		xit("should remove the mask only when the config is set to false", () => {
-			changeInputValue(inputElement, "my-email@");
+		// FIXME NG0100: ExpressionChangedAfterItHasBeenCheckedError - #2860 https://github.com/NationalBankBelgium/stark/issues/2860
+		xit("should refresh the mask whenever the configuration changes", () => {
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			expect(hostComponent.ngModelValue).toBe("my-email@ .");
+			expect(hostComponent.ngModelValue).toBe("12,345");
 
-			hostComponent.emailMaskConfig = undefined;
+			hostComponent.numberMaskConfig = {
+				...numberMaskConfig,
+				prefix: "%",
+				suffix: " percent",
+				thousandsSeparatorSymbol: "-"
+			};
 			fixture.detectChanges();
 
-			changeInputValue(inputElement, "what@.ever@.");
+			expect(hostComponent.ngModelValue).toBe("%12-345 percent");
+		});
+
+		xit("should remove the mask when the config is undefined", () => {
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			expect(hostComponent.ngModelValue).toBe("my-email@ ."); // the mask is enabled by default
+			expect(hostComponent.ngModelValue).toBe("12,345");
 
-			hostComponent.emailMaskConfig = ""; // use case when the directive is used with no inputs: <input type='text' [(ngModel)]='ngModelValue' starkEmailMask>
+			hostComponent.numberMaskConfig = <any>undefined;
 			fixture.detectChanges();
 
-			changeInputValue(inputElement, "what@.ever@.");
+			changeInputValue(inputElement, "whatever+1*23-4/5");
 			fixture.detectChanges();
 
-			expect(hostComponent.ngModelValue).toBe("my-email@ ."); // the mask is enabled by default
-
-			hostComponent.emailMaskConfig = false;
-			fixture.detectChanges();
-
-			changeInputValue(inputElement, "what@@.ever@.");
-			fixture.detectChanges();
-
-			expect(hostComponent.ngModelValue).toBe("what@@.ever@."); // no mask at all
+			expect(hostComponent.ngModelValue).toBe("whatever+1*23-4/5"); // no mask at all
 		});
 	});
 
@@ -240,7 +269,7 @@ describe("EmailMaskDirective", () => {
 
 		beforeEach(
 			waitForAsync(() => {
-				const newTemplate: string = getTemplate("[formControl]='formControl' [starkEmailMaskNew]='emailMaskConfig'");
+				const newTemplate: string = getTemplate("[formControl]='formControl' [starkNumberMaskNew]='numberMaskConfig'");
 
 				TestBed.overrideTemplate(TestComponent, newTemplate);
 
@@ -257,7 +286,7 @@ describe("EmailMaskDirective", () => {
 		});
 
 		it("should render the appropriate content", () => {
-			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkEmailMask directive
+			expect(inputElement.attributes["ng-reflect-mask-config"]).toBeDefined(); // starkNumberMask directive
 		});
 
 		xit("should update the input value and show the mask only when a valid event is triggered in the input field", () => {
@@ -272,10 +301,10 @@ describe("EmailMaskDirective", () => {
 				expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
 
 				mockValueChangeObserver.next.calls.reset();
-				changeInputValue(inputElement, "my-email@", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
-				expect(hostComponent.formControl.value).toBe("my-email@ .");
+				expect(hostComponent.formControl.value).toBe("12,345");
 				// FIXME Check why it is called twice instead of once
 				expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
 				expect(mockValueChangeObserver.error).not.toHaveBeenCalled();
@@ -293,7 +322,7 @@ describe("EmailMaskDirective", () => {
 				expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
 
 				mockValueChangeObserver.next.calls.reset();
-				changeInputValue(inputElement, "my-email@", eventType);
+				changeInputValue(inputElement, "12345", eventType);
 				fixture.detectChanges();
 
 				// IMPORTANT: the formControl is not changed with invalid events, just with "input" events
@@ -305,7 +334,7 @@ describe("EmailMaskDirective", () => {
 		});
 
 		xit("should prevent invalid values to be entered in the input field when the value is changed manually", () => {
-			const invalidValues: string[] = ["@@", "@.a.", " @ .", "what@.ever@."];
+			const invalidValues: string[] = ["a", " ", "/*-+,."];
 
 			for (const value of invalidValues) {
 				mockValueChangeObserver.next.calls.reset();
@@ -318,52 +347,64 @@ describe("EmailMaskDirective", () => {
 				expect(mockValueChangeObserver.error).not.toHaveBeenCalled();
 				expect(mockValueChangeObserver.complete).not.toHaveBeenCalled();
 			}
+
+			const invalidNumericValues: string[] = ["1-2-3.4.5", "+1*23-4/5", ".1.234,5"];
+
+			for (const numericValue of invalidNumericValues) {
+				mockValueChangeObserver.next.calls.reset();
+				changeInputValue(inputElement, numericValue);
+				fixture.detectChanges();
+
+				expect(hostComponent.formControl.value).toBe("12,345");
+				// FIXME Check why it is called twice instead of once
+				expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
+				expect(mockValueChangeObserver.error).not.toHaveBeenCalled();
+				expect(mockValueChangeObserver.complete).not.toHaveBeenCalled();
+			}
+		});
+
+		xit("should refresh the mask whenever the configuration changes", () => {
+			changeInputValue(inputElement, "12345");
+			fixture.detectChanges();
+
+			expect(hostComponent.formControl.value).toBe("12,345");
+			// FIXME Check why it is called twice instead of once
+			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
+
+			mockValueChangeObserver.next.calls.reset();
+			hostComponent.numberMaskConfig = {
+				...numberMaskConfig,
+				prefix: "%",
+				suffix: " percent",
+				thousandsSeparatorSymbol: "-"
+			};
+			fixture.detectChanges();
+
+			expect(hostComponent.formControl.value).toBe("%12-345 percent");
+			// FIXME Check why it is called twice instead of once
+			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
+			expect(mockValueChangeObserver.error).not.toHaveBeenCalled();
+			expect(mockValueChangeObserver.complete).not.toHaveBeenCalled();
 		});
 
 		xit("should remove the mask when the config is undefined", () => {
-			changeInputValue(inputElement, "my-email@");
+			changeInputValue(inputElement, "12345");
 			fixture.detectChanges();
 
-			expect(hostComponent.formControl.value).toBe("my-email@ .");
+			expect(hostComponent.formControl.value).toBe("12,345");
 			// FIXME Check why it is called twice instead of once
 			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
 
 			mockValueChangeObserver.next.calls.reset();
-			hostComponent.emailMaskConfig = undefined;
-			fixture.detectChanges();
-			expect(mockValueChangeObserver.next).not.toHaveBeenCalled(); // no value change, the mask is enabled by default
-
-			mockValueChangeObserver.next.calls.reset();
-			changeInputValue(inputElement, "what@.ever@.");
-			fixture.detectChanges();
-
-			expect(hostComponent.formControl.value).toBe("my-email@ ."); // the mask is enabled by default
-			// FIXME Check why it is called twice instead of once
-			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
-
-			mockValueChangeObserver.next.calls.reset();
-			hostComponent.emailMaskConfig = ""; // use case when the directive is used with no inputs: <input type='text' [formControl]='formControl' starkEmailMask>
-			fixture.detectChanges();
-			expect(mockValueChangeObserver.next).not.toHaveBeenCalled(); // no value change, the mask is enabled by default
-
-			mockValueChangeObserver.next.calls.reset();
-			changeInputValue(inputElement, "what@.ever@.");
-			fixture.detectChanges();
-
-			expect(hostComponent.formControl.value).toBe("my-email@ ."); // the mask is enabled by default
-			// FIXME Check why it is called twice instead of once
-			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
-
-			mockValueChangeObserver.next.calls.reset();
-			hostComponent.emailMaskConfig = false;
+			hostComponent.numberMaskConfig = <any>undefined;
 			fixture.detectChanges();
 			expect(mockValueChangeObserver.next).not.toHaveBeenCalled(); // no value change, the mask was just disabled
 
 			mockValueChangeObserver.next.calls.reset();
-			changeInputValue(inputElement, "what@@.ever@.");
+			changeInputValue(inputElement, "whatever+1*23-4/5");
 			fixture.detectChanges();
 
-			expect(hostComponent.formControl.value).toBe("what@@.ever@."); // no mask at all
+			expect(hostComponent.formControl.value).toBe("whatever+1*23-4/5"); // no mask at all
 			// FIXME Check why it is called twice instead of once
 			expect(mockValueChangeObserver.next).toHaveBeenCalledTimes(2);
 			expect(mockValueChangeObserver.error).not.toHaveBeenCalled();
