+++ packages/stark-ui/src/modules/input-mask-directives/directives/text-mask-config.intf.ts
