+++ packages/stark-ui/src/modules/input-mask-directives/directives/timestamp-mask.directive.spec.ts
